backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.4.2
supabase==2.3.0
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.13
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
numpy==1.26.2
faiss-cpu==1.7.4
